Inspector: share the DAG canvas — RunTreeDock onto @corpus/ui/RunTreeCanvas
The dock embedded its own ~220-line copy of the build-provenance DAG canvas
(SVG edges + node cards + lane layout + the incident-edge focus highlight) —
byte-identical to the shared @corpus/ui/RunTreeDag's canvas, the duplication the
dock's own code comment flagged ("fold both onto a shared component once that
settles"). Extract that canvas into a new shared @corpus/ui/RunTreeCanvas and
point BOTH surfaces at it:

- @corpus/ui/RunTreeCanvas — presentational, controlled (tree + selected /
  expanded via props). Renders only the canvas (loading / empty / the positioned
  layout with edges + node cards + click-to-select + the per-node expand button);
  the host supplies the surrounding scroll box and whatever node-detail pane it
  wants. SVG marker ids are now per-instance (createUniqueId) so two DAGs mounted
  in one document can't cross-resolve their arrowheads.
- RunTreeDag is now canvas + detail-slot around RunTreeCanvas (no visual change).
- RunTreeDock keeps its richer node-detail pane, FreshnessPanel, resize, the
  rootInstances picker, the waterfall, and the by-anchor groups — only its inline
  DAG canvas is replaced by <RunTreeCanvas>; the dead canvas helpers
  (layout/connected/isIncident/nodeY/hasKids/badgeColor) and their imports are
  removed.

Net: one DAG renderer behind talmud's dock, talmud's AlignPage, and the tanach
inspector + alignment page. 113 core + 64 tanach + 2551 talmud tests green;
typecheck + lint + both builds clean. Reviewed read-only with Codex (dock
selection / expand / focus-highlight / detail / freshness / resize / rootInstances
all preserved, no dangling refs; the marker-id collision it flagged is fixed).
