+++ packages/talmud/src/client/RunTreeDock.tsx
@@ -18,35 +18,24 @@
 
 import {
   ACTIVE_STROKE,
-  AuthorityBadge,
   BADGE_LLM,
   BADGE_PRO,
   BADGE_SRC,
   CANVAS,
-  CANVAS_BORDER,
-  CARD_STROKE,
-  computeLayout,
   displayLabel,
-  edgePath,
   fmtCost,
   fmtMs,
   type IconVariant,
-  type LaidEdge,
-  type Layout,
-  LEFT_PAD,
-  NODE_H,
-  NODE_W,
   NodeIcon,
   ProvenanceSection,
-  ROW_H,
   type RunResult,
   type RunTree,
   STALENESS_COLOR,
   StalenessDot,
-  TOP_PAD,
   type TreeNode,
   variantOf,
 } from '@corpus/ui/RunTree';
+import { RunTreeCanvas } from '@corpus/ui/RunTreeCanvas';
 import {
   createEffect,
   createMemo,
@@ -783,35 +772,17 @@ export default function RunTreeDock(props: {
     },
   );
 
-  const layout = createMemo<Layout | null>(() => {
-    const t = tree();
-    return t ? computeLayout(t, expanded()) : null;
-  });
+  // The DAG canvas (layout, edges, node cards, the incident-edge focus
+  // highlight) is the shared @corpus/ui/RunTreeCanvas; the dock keeps only the
+  // waterfall + its richer node-detail/freshness panes around it. nodeOf +
+  // toggleExpand remain — the detail pane and the canvas props use them.
   const nodeOf = (id: string): TreeNode | undefined => tree()?.nodes[id];
-  const hasKids = (id: string): boolean => !!tree()?.edges.some((e) => e[0] === id);
   const toggleExpand = (id: string) =>
     setExpanded((prev) => {
       const n = new Set(prev);
       n.has(id) ? n.delete(id) : n.add(id);
       return n;
     });
-  const badgeColor = (n: TreeNode) =>
-    n.kind !== 'llm' ? BADGE_SRC : n.model?.includes('pro') ? BADGE_PRO : BADGE_LLM;
-
-  // The selected node + everything one edge away — for the focus highlight
-  // (incident edges drawn bold, the rest faded).
-  const connected = createMemo<Set<string>>(() => {
-    const sel = selected();
-    const lay = layout();
-    if (!sel || !lay) return new Set();
-    const set = new Set<string>([sel]);
-    for (const e of lay.edges) {
-      if (e.fromId === sel) set.add(e.toId);
-      if (e.toId === sel) set.add(e.fromId);
-    }
-    return set;
-  });
-  const isIncident = (e: LaidEdge) => e.fromId === selected() || e.toId === selected();
 
   const [detail] = createResource(
     () => {
@@ -873,11 +844,6 @@ export default function RunTreeDock(props: {
     };
     document.addEventListener('mousemove', move);
     document.addEventListener('mouseup', up);
-  };
-
-  const nodeY = (id: string) => {
-    const r = layout()!.rowOf.get(id)!;
-    return TOP_PAD + r * ROW_H;
   };
 
   // Push the daf left by the panel width when open (mirrors the old left shelf).
@@ -1169,220 +1135,14 @@ export default function RunTreeDock(props: {
               padding: '0.5rem',
             }}
           >
-            <Show when={tree.loading}>
-              <div style={{ padding: '0.5rem', color: '#aaa' }}>loading…</div>
-            </Show>
-            <Show when={tree() === null && !tree.loading}>
-              <div style={{ padding: '0.5rem', color: '#c00' }}>
-                no graph (unknown piece, or nothing cached)
-              </div>
-            </Show>
-            <Show when={layout()}>
-              {(lay) => (
-                <div
-                  style={{
-                    position: 'relative',
-                    width: `${lay().width}px`,
-                    height: `${lay().height}px`,
-                    border: `1px solid ${CANVAS_BORDER}`,
-                    'border-radius': '8px',
-                    background: CANVAS,
-                  }}
-                >
-                  {/* edge layer */}
-                  <svg
-                    aria-hidden="true"
-                    width={lay().width}
-                    height={lay().height}
-                    style={{
-                      position: 'absolute',
-                      inset: 0,
-                      'pointer-events': 'none',
-                      overflow: 'visible',
-                    }}
-                  >
-                    <defs>
-                      <marker
-                        id="rt-arrow"
-                        markerWidth="8"
-                        markerHeight="8"
-                        refX="6"
-                        refY="3"
-                        orient="auto"
-                      >
-                        <path d="M0 0 L6 3 L0 6 z" fill="#c9b8b0" />
-                      </marker>
-                      <marker
-                        id="rt-arrow-hot"
-                        markerWidth="8"
-                        markerHeight="8"
-                        refX="6"
-                        refY="3"
-                        orient="auto"
-                      >
-                        <path d="M0 0 L6 3 L0 6 z" fill="#8a2a2b" />
-                      </marker>
-                    </defs>
-                    <For each={lay().edges}>
-                      {(e) => {
-                        const hot = () => isIncident(e);
-                        const faded = () => !!selected() && !hot();
-                        return (
-                          // arrow points dependency -> consumer ("what feeds into what")
-                          <path
-                            d={edgePath(e.toRow, e.fromRow, e.lane)}
-                            fill="none"
-                            stroke={hot() ? '#8a2a2b' : '#d3c4ba'}
-                            stroke-width={hot() ? 2 : 1.5}
-                            stroke-opacity={faded() ? 0.22 : hot() ? 0.85 : 1}
-                            stroke-linecap="round"
-                            stroke-linejoin="round"
-                            marker-end={`url(#${hot() ? 'rt-arrow-hot' : 'rt-arrow'})`}
-                          />
-                        );
-                      }}
-                    </For>
-                  </svg>
-                  {/* node cards */}
-                  <For each={lay().order}>
-                    {(id) => {
-                      const n = () => nodeOf(id)!;
-                      const isLLM = () => n().kind === 'llm';
-                      const sel = () => selected() === id;
-                      const exp = () => expanded().has(id);
-                      const slow = () => (n().cold_ms ?? 0) > 10_000;
-                      const dim = () => !!selected() && !connected().has(id);
-                      const activate = () => {
-                        setSelected(id);
-                        if (hasKids(id)) toggleExpand(id);
-                      };
-                      return (
-                        // biome-ignore lint/a11y/useSemanticElements: node card contains a nested expand <button>; a native button cannot contain another button
-                        <div
-                          role="button"
-                          tabIndex={0}
-                          onClick={activate}
-                          onKeyDown={(e) => {
-                            if (e.key === 'Enter' || e.key === ' ') {
-                              e.preventDefault();
-                              activate();
-                            }
-                          }}
-                          style={{
-                            position: 'absolute',
-                            left: `${LEFT_PAD}px`,
-                            top: `${nodeY(id)}px`,
-                            width: `${NODE_W}px`,
-                            height: `${NODE_H}px`,
-                            display: 'flex',
-                            'align-items': 'center',
-                            gap: '0.5rem',
-                            padding: '0 0.6rem',
-                            cursor: 'pointer',
-                            'box-sizing': 'border-box',
-                            background: sel() ? '#fdf2f2' : '#fff',
-                            border: `${sel() ? 1.75 : 1}px solid ${sel() ? ACTIVE_STROKE : CARD_STROKE}`,
-                            'border-radius': '11px',
-                            'box-shadow': '0 1px 2px rgba(58,51,32,0.08)',
-                            opacity: dim() ? 0.42 : 1,
-                            transition: 'opacity 0.12s',
-                          }}
-                        >
-                          <NodeIcon variant={variantOf(n())} color={badgeColor(n())} />
-                          <div style={{ flex: 1, 'min-width': 0 }}>
-                            <div
-                              style={{ display: 'flex', 'align-items': 'baseline', gap: '0.4rem' }}
-                            >
-                              <span
-                                style={{
-                                  'font-weight': 600,
-                                  'font-size': '0.84rem',
-                                  color: '#2a2723',
-                                  'white-space': 'nowrap',
-                                  overflow: 'hidden',
-                                  'text-overflow': 'ellipsis',
-                                }}
-                              >
-                                {displayLabel(n().id, n().label)}
-                              </span>
-                              <span
-                                style={{
-                                  'margin-left': 'auto',
-                                  'font-size': '0.68rem',
-                                  'font-variant-numeric': 'tabular-nums',
-                                  color: slow() ? '#b45309' : '#9a857c',
-                                  'flex-shrink': 0,
-                                }}
-                              >
-                                {fmtMs(n().cold_ms)}
-                              </span>
-                              <Show when={n().staleness}>
-                                {(s) => (
-                                  <StalenessDot
-                                    staleness={s()}
-                                    inputsChanged={n().inputsChanged}
-                                    isMark={n().producer === 'mark'}
-                                  />
-                                )}
-                              </Show>
-                            </div>
-                            <div
-                              style={{
-                                display: 'flex',
-                                'align-items': 'center',
-                                gap: '0.3rem',
-                                'font-size': '0.66rem',
-                                'font-family': 'ui-monospace, Menlo, monospace',
-                                color: isLLM() ? '#9a8fb5' : '#9aa4ad',
-                                'white-space': 'nowrap',
-                                overflow: 'hidden',
-                                'text-overflow': 'ellipsis',
-                              }}
-                            >
-                              <Show when={n().authority}>
-                                {(a) => <AuthorityBadge authority={a()} />}
-                              </Show>
-                              {isLLM()
-                                ? `${(n().model ?? '').split('/').pop()} · ${fmtCost(n().cost)}`
-                                : 'source · $0'}
-                            </div>
-                          </div>
-                          <Show when={hasKids(id)}>
-                            <button
-                              type="button"
-                              onClick={(ev) => {
-                                ev.stopPropagation();
-                                setSelected(id);
-                                toggleExpand(id);
-                              }}
-                              title={exp() ? 'collapse inputs' : 'expand inputs'}
-                              style={{
-                                'flex-shrink': 0,
-                                width: '18px',
-                                height: '18px',
-                                'border-radius': '50%',
-                                border: '1px solid #d8c9c0',
-                                background: '#fff',
-                                color: '#8a7d74',
-                                cursor: 'pointer',
-                                'font-size': '0.8rem',
-                                'line-height': 1,
-                                display: 'inline-flex',
-                                'align-items': 'center',
-                                'justify-content': 'center',
-                                padding: 0,
-                              }}
-                            >
-                              {exp() ? '–' : '+'}
-                            </button>
-                          </Show>
-                        </div>
-                      );
-                    }}
-                  </For>
-                </div>
-              )}
-            </Show>
+            <RunTreeCanvas
+              tree={tree() ?? null}
+              loading={tree.loading}
+              selected={selected()}
+              onSelect={setSelected}
+              expanded={expanded()}
+              onToggleExpand={toggleExpand}
+            />
           </div>
         </Show>
