+++ packages/ui/package.json
@@ -29,6 +29,7 @@
     "./LoadProgress": "./src/LoadProgress.tsx",
     "./InspectorRow": "./src/InspectorRow.tsx",
     "./RunTree": "./src/RunTree.tsx",
+    "./RunTreeCanvas": "./src/RunTreeCanvas.tsx",
     "./RunTreeDag": "./src/RunTreeDag.tsx",
     "./RunWaterfall": "./src/RunWaterfall.tsx",
     "./UsagePage": "./src/UsagePage.tsx"
